fix: attrNames.count Int→UInt（v0.3.43 重打）

+++ EscapeOS/Views/ProcessManagerView.swift
@@ -395,7 +395,7 @@ final class ProcessManagerService {
             var config = IdeviceSysmontapConfig(
                 interval_ms: 500,
                 process_attributes: attrPtrs,
-                process_attributes_count: attrNames.count,
+                process_attributes_count: UInt(attrNames.count),
                 system_attributes: nil,
                 system_attributes_count: 0
             )
